+++ test/test_aeon2yw_.py
@@ -14,6 +14,7 @@ import zipfile
 import codecs
 import json
 import aeon2yw_
+from json import JSONDecodeError
 
 # Test environment
 
@@ -36,22 +37,10 @@ def open_timeline(filePath):
     """Unzip the project file and read 'timeline.json'.
     Return the JSON timeline structure.
     """
-    try:
-        with zipfile.ZipFile(filePath, 'r') as myzip:
-            jsonBytes = myzip.read('timeline.json')
-            jsonStr = codecs.decode(jsonBytes, encoding='utf-8')
-    except:
-        raise Error(f'Cannot read JSON data.')
-
-    if not jsonStr:
-        raise Error(f'No JSON part found.')
-
-    try:
-        jsonData = json.loads(jsonStr)
-    except('JSONDecodeError'):
-        raise Error(f'Invalid JSON data.')
-
-    return jsonData
+    with zipfile.ZipFile(filePath, 'r') as myzip:
+        jsonBytes = myzip.read('timeline.json')
+        jsonStr = codecs.decode(jsonBytes, encoding='utf-8')
+    return json.loads(jsonStr)
 
 
 def read_file(inputFile):
@@ -103,20 +92,22 @@ class NormalOperation(unittest.TestCase):
             pass
         remove_all_testfiles()
 
+    # @unittest.skip('')
     def test_ambiguous_aeon_event(self):
-        copyfile(TEST_DATA_PATH + 'update_notes.ini', INI_FILE)
         copyfile(TEST_DATA_PATH + 'normal.aeonzip', TEST_AEON)
         os.chdir(TEST_EXEC_PATH)
         aeon2yw_.run(TEST_AEON, silentMode=True)
         self.assertStderrEquals('FAIL: Ambiguous Aeon event title "Mrs Hubbard sleeps".')
 
+    # @unittest.skip('')
     def test_create_yw7(self):
-        copyfile(TEST_DATA_PATH + 'update_notes.ini', INI_FILE)
+        copyfile(TEST_DATA_PATH + 'create_notes.ini', INI_FILE)
         copyfile(TEST_DATA_PATH + 'date_limits.aeonzip', TEST_AEON)
         os.chdir(TEST_EXEC_PATH)
         aeon2yw_.run(TEST_AEON, silentMode=True)
         self.assertEqual(read_file(TEST_YW7), read_file(TEST_DATA_PATH + 'date_limits_notes.yw7'))
 
+    # @unittest.skip('')
     def test_create_yw7_narrative_only(self):
         copyfile(TEST_DATA_PATH + 'narrative_only.ini', INI_FILE)
         copyfile(TEST_DATA_PATH + 'date_limits.aeonzip', TEST_AEON)
@@ -124,24 +115,27 @@ class NormalOperation(unittest.TestCase):
         aeon2yw_.run(TEST_AEON, silentMode=True)
         self.assertEqual(read_file(TEST_YW7), read_file(TEST_DATA_PATH + 'date_limits.yw7'))
 
+    # @unittest.skip('')
     def test_update_yw7(self):
-        copyfile(TEST_DATA_PATH + 'update_notes.ini', INI_FILE)
+        copyfile(TEST_DATA_PATH + 'create_notes.ini', INI_FILE)
         copyfile(TEST_DATA_PATH + 'date_limits.yw7', TEST_YW7)
         copyfile(TEST_DATA_PATH + 'updated.aeonzip', TEST_AEON)
         os.chdir(TEST_EXEC_PATH)
         aeon2yw_.run(TEST_AEON, silentMode=True)
-        self.assertEqual(read_file(TEST_YW7), read_file(TEST_DATA_PATH + 'updated_notes.yw7'))
+        self.assertEqual(read_file(TEST_YW7), read_file(TEST_DATA_PATH + 'updated_from_aeon_notes.yw7'))
         self.assertEqual(read_file(TEST_YW7_BAK), read_file(TEST_DATA_PATH + 'date_limits.yw7'))
 
+    # @unittest.skip('')
     def test_update_yw7_narrative_only(self):
         copyfile(TEST_DATA_PATH + 'narrative_only.ini', INI_FILE)
         copyfile(TEST_DATA_PATH + 'date_limits.yw7', TEST_YW7)
         copyfile(TEST_DATA_PATH + 'updated.aeonzip', TEST_AEON)
         os.chdir(TEST_EXEC_PATH)
         aeon2yw_.run(TEST_AEON, silentMode=True)
-        self.assertEqual(read_file(TEST_YW7), read_file(TEST_DATA_PATH + 'updated.yw7'))
+        self.assertEqual(read_file(TEST_YW7), read_file(TEST_DATA_PATH + 'updated_from_aeon.yw7'))
         self.assertEqual(read_file(TEST_YW7_BAK), read_file(TEST_DATA_PATH + 'date_limits.yw7'))
 
+    # @unittest.skip('')
     def test_create_date_limits_aeon(self):
         copyfile(TEST_DATA_PATH + 'date_limits.yw7', TEST_YW7)
         copyfile(TEST_DATA_PATH + 'minimal.aeonzip', TEST_AEON)
@@ -149,12 +143,29 @@ class NormalOperation(unittest.TestCase):
         aeon2yw_.run(TEST_YW7, silentMode=True)
         self.assertEqual(open_timeline(TEST_AEON), open_timeline(TEST_DATA_PATH + 'created.aeonzip'))
 
+    # @unittest.skip('')
     def test_create_arc_aeon(self):
         copyfile(TEST_DATA_PATH + 'arc.yw7', TEST_YW7)
         copyfile(TEST_DATA_PATH + 'minimal.aeonzip', TEST_AEON)
         os.chdir(TEST_EXEC_PATH)
         aeon2yw_.run(TEST_YW7, silentMode=True)
         self.assertEqual(open_timeline(TEST_AEON), open_timeline(TEST_DATA_PATH + 'created_arc.aeonzip'))
+
+    # @unittest.skip('')
+    def test_update_aeon(self):
+        copyfile(TEST_DATA_PATH + 'updated.yw7', TEST_YW7)
+        copyfile(TEST_DATA_PATH + 'created.aeonzip', TEST_AEON)
+        os.chdir(TEST_EXEC_PATH)
+        aeon2yw_.run(TEST_YW7, silentMode=True)
+        self.assertEqual(open_timeline(TEST_AEON), open_timeline(TEST_DATA_PATH + 'updated_from_yw.aeonzip'))
+
+    # @unittest.skip('')
+    def test_update1_aeon(self):
+        copyfile(TEST_DATA_PATH + 'updated1.yw7', TEST_YW7)
+        copyfile(TEST_DATA_PATH + 'updated_from_yw.aeonzip', TEST_AEON)
+        os.chdir(TEST_EXEC_PATH)
+        aeon2yw_.run(TEST_YW7, silentMode=True)
+        self.assertEqual(open_timeline(TEST_AEON), open_timeline(TEST_DATA_PATH + 'updated1_from_yw.aeonzip'))
 
     def tearDown(self):
         sys.stdout = self.original_output